test(redteam): extract vfr-rt audit spec + close complete_empty gap + non-vacuity guards
- extract VFR-RT audit describe (started/completed/expired) from audit-completeness.spec.ts
  (583->458L, back under the 500 test-file ceiling) into audit-completeness-vfr-rt.spec.ts
- add 4th audit test: vfr_rt_exam.completed via complete_empty_exam_session on an empty
  non-overdue vfr session (closes red-team gap; mig 102 L264 non-overdue branch)
- assert answers.length === questions.length at buildVfrRtAnswers call sites (semantic-reviewer)
- bump red-team spec count 51->52 (tech.md x3, decisions.md)
