+++ apps/web/e2e/redteam/rpc-vfr-rt-results.spec.ts
@@ -270,6 +270,9 @@ test.describe('Red Team: get_vfr_rt_exam_results RPC — success / output contra
     expect(questions.length).toBe(25)
 
     const answers = buildVfrRtAnswers(questions, { failPart2: opts.failPart2 })
+    // Non-vacuous: one answer per question — guards a silent unknown-type skip
+    // in the helper that would otherwise under-count the graded payload.
+    expect(answers.length).toBe(questions.length)
     const { data: submitRaw, error: submitErr } = await victimClient.rpc(
       'submit_vfr_rt_exam_answers',
       { p_session_id: started.session_id, p_answers: answers },
