+++ docs/decisions.md
@@ -478,7 +478,7 @@ Full audit completed — 46 files reviewed. Score: 9.5/10. Full report: `docs/se
 **Context:** Static code review and unit tests miss real exploit chains and race conditions. Need active adversarial testing against a running app to prove defenses hold.
 
 **Decided:**
-- Create red-team suite: Playwright attack specs executing exploit chains against local Supabase (same as production); 51 specs as of 2026-06-29 covering OWASP A01/A02/A03/A07/A09 and A10:2025 (exceptional-condition handling / error-path information disclosure)
+- Create red-team suite: Playwright attack specs executing exploit chains against local Supabase (same as production); 52 specs as of 2026-07-03 covering OWASP A01/A02/A03/A07/A09 and A10:2025 (exceptional-condition handling / error-path information disclosure)
 - Attack vectors cover: RLS bypass (cross-tenant, question membership), RPC boundary breaches, session forgery (PKCE, replay), race conditions (concurrent discard+complete), audit log tampering and completeness, quiz draft injection, SQL/XSS injection, security-header validation
 - Separate Playwright project (`e2e/redteam/`) to avoid clutter, testIgnore on normal e2e pipeline
 - Red-team agent (sonnet) triggers post-commit on security-sensitive file changes (migrations, db/src, quiz/actions, auth, proxy.ts, security.md) — maps diff to affected specs, flags coverage gaps
